+++ include/natalie/types.hpp
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <limits.h>
+#include <stdint.h>
 
 #define NAT_INT_MIN LLONG_MIN
 #define NAT_INT_MAX LLONG_MAX
 
 namespace Natalie {
 
-using nat_int_t = long long;
+using nat_int_t = int64_t;
 
 }
